sched/fair: Add fallback CPU selection
When all CPUs are overloaded, the task is placed on the previous
CPU during wakeup. This can result in packing high number of tasks
on a CPU. When no suitable CPU is found and previous CPU has more
than 32 runnable tasks, add a fallback CPU selection to find
the CPU with least number of runnable tasks.

When a CPU is isolated or hotplugged out, all tasks queued on
the CPU are migrated to CPU#0. This can result in packing high
number of tasks on CPU#0. Add the above fallback CPU selection
here also.

eng.stk: adapt for OP patchset

Change-Id: Ia9b959339692770b4cc594fa36cf42ca342cf2de
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>